Adding a switch to the main menu to allow the switching between the real and mocked out feature flag service.
